remove obsolete OpenStack4 support
