+++ lib/iaas/iaas.rb
@@ -37,7 +37,7 @@ module BushSlicer
         iaas_conf_params[params[0].strip] = params[1].strip
       end
 
-      return BushSlicer::OpenStack.instance(
+      return BushSlicer::OpenStack.new(
         :url => iaas_conf_params['auth-url'] + "auth/tokens",
         :user => iaas_conf_params["username"],
         :password => iaas_conf_params["password"],
